+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 - Added sphinx docu.
+- Added CLI command 'identify'
 
 ## v0.1.0 - 29.01.24
 - Initial release, based on [https://gitlab.com/pyshacks/pnio_dcp](https://gitlab.com/pyshacks/pnio_dcp) version 1.2.